# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MEV Arbitrage Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List executed trades
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, failed, reverted]
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
  /wallet/connect:
    post:
      operationId: connectWallet
      tags: [wallet]
      summary: Connect a wallet using private key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectWalletBody"
      responses:
        "200":
          description: Wallet connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletStatus"
        "400":
          description: Invalid private key or connection error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /execute/{opportunityId}:
    post:
      operationId: executeOpportunity
      tags: [execute]
      summary: Execute an arbitrage opportunity
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          description: Execution error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        opportunityId:
          type: integer
        txHash:
          type: string
        strategy:
          type: string
        tokenPair:
          type: string
        amountIn:
          type: string
        amountOut:
          type: string
        profit:
          type: string
        profitUsd:
          type: string
        gasUsed:
          type: string
        gasCost:
          type: string
        gasCostUsd:
          type: string
        netProfitUsd:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed, reverted]
        blockNumber:
          type: integer
        executedAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
        flashLoanAmount:
          type: string
        flashLoanFee:
          type: string
        buyDex:
          type: string
        sellDex:
          type: string
        error:
          type: string
      required:
        - id
        - strategy
        - tokenPair
        - status
        - executedAt
...
    WalletStatus:
      type: object
      properties:
        connected:
          type: boolean
        address:
          type: string
        balance:
          type: string
          description: ETH/MATIC balance
        network:
          type: string
        contractAddress:
          type: string
        contractValid:
          type: boolean
        contractOwner:
          type: string
      required:
        - connected
    ConnectWalletBody:
      type: object
      properties:
        privateKey:
          type: string
          description: Ethereum private key (hex)
        contractAddress:
          type: string
          description: Deployed FlashLoanArbitrage contract address
      required:
        - privateKey
    ExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        txHash:
          type: string
        blockNumber:
          type: integer
        gasUsed:
          type: string
        gasCostEth:
          type: string
        gasCostUsd:
          type: string
        profitEth:
          type: string
        profitUsd:
          type: string
        error:
          type: string
        simulated:
          type: boolean
        opportunityId:
          type: integer
      required:
        - success
        - simulated
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error